+++ .lintstagedrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  '*.{tsx,ts}': ['eslint --cache --fix --max-warnings=0'],
+};