+++ src/ExpoMaterial3ThemeModule.ts
@@ -1,5 +1,7 @@
 import { requireNativeModule } from 'expo-modules-core';
 
-// It loads the native module object from the JSI or falls back to
-// the bridge module (from NativeModulesProxy) if the remote debugger is on.
-export default requireNativeModule('ExpoMaterial3Theme');
+let ExpoMaterial3ThemeModule;
+try {
+  ExpoMaterial3ThemeModule = requireNativeModule('ExpoMaterial3Theme');
+} catch {}
+export default ExpoMaterial3ThemeModule;