+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [1.0.0-beta.2](https://github.com/pchmn/expo-material3-theme/compare/v1.0.0-beta.1...v1.0.0-beta.2) (2023-03-30)
+
+
+### Features
+
+* make module available for Expo Go ([a6cf7d1](https://github.com/pchmn/expo-material3-theme/commit/a6cf7d1f7b6a29a91faf902304490255626bcaa0))
+* update `useMaterial3Theme` params ([3329dd4](https://github.com/pchmn/expo-material3-theme/commit/3329dd4718f4d0bf2b88fe83d74380f208c5ad53))
+
+
+### BREAKING CHANGES
+
+* remove `overwrite` param
+
+`overwrite` param in `useMaterial3Theme` is removed, use `sourceColor` param to overwrite system theme
+
 # 1.0.0-beta.1 (2023-03-16)
 
 First version